move cirun stuff to cirun utils
