+++ conda_smithy/cli.py
@@ -12,6 +12,8 @@ from textwrap import dedent
 import conda  # noqa
 import conda_build.api
 from conda_build.metadata import MetaData
+
+import conda_smithy.cirun_utils
 from conda_smithy.utils import get_feedstock_name_from_meta, merge_dict
 from ruamel.yaml import YAML
 
@@ -299,7 +301,7 @@ class RegisterCI(Subcommand):
             args.anaconda_token,
         ]):
             raise RuntimeError(
-            "The --remove flag is only supported for Cirun for now"
+                "The --remove flag is only supported for Cirun for now"
             )
         if not args.anaconda_token:
             print(
@@ -377,7 +379,7 @@ class RegisterCI(Subcommand):
                     to_remove = ["*"]
 
                 for resource in to_remove:
-                    ci_register.remove_project_from_cirun_resource(
+                    conda_smithy.cirun_utils.remove_project_from_cirun_resource(
                         owner, repo, resource
                     )
                 # current_resources = ci_register.enabled_cirun_resources(owner, repo)
@@ -385,9 +387,9 @@ class RegisterCI(Subcommand):
                 #     ci_register.remove_project_from_cirun(owner, repo)
             else:
                 # ci_register.ensure_cirun_app_installed(owner, repo)
-                ci_register.enable_cirun_for_project(owner, repo)
+                conda_smithy.cirun_utils.enable_cirun_for_project(owner, repo)
                 for resource in args.cirun_resources:
-                    ci_register.add_project_to_cirun_resource(owner, repo, resource)
+                    conda_smithy.cirun_utils.add_project_to_cirun_resource(owner, repo, resource)
         else:
             print("Cirun registration disabled.")
